[JENKINS-28689] Tests has been reviewed. Clean up source code
